renamed folders and relinked
