+++ client/src/components/Featured.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 // Styles 
-import "../styles/featComp.scss";
+import "../sass/components/featComp.scss";
 
 const Featured = () => {
   return (
