+++ client/src/sass/components/aboutComp.scss
