+++ client/src/sass/components/featComp.scss
