+++ client/src/sass/components/footerComp.scss
