+++ client/src/sass/components/heroComp.scss
